Move config and runtime to CUE
As I looked at the combinatorial challenge of creating examples, and
catering for users, that use tools with subtly different and unique
file-name forms, I realised this tool would eventually (probably soon!)
have to allow the user to indicate some kind of
text-processing/templating/replacement in the config file.

After a bit of playing around with sed, its `--sandbox` safety net, and
so on, I realised that building my own system for that would be a
nightmare.

This was the nudge I needed to move the config to live as .cue, as CUE
has just such a templating system built in to its very core. Plus, CUE
is far nicer to automate around, and has string-manipulation available
for free. E.g. the current `go.arch` default in `bin/facts.cue` in this
commit.

Introducing a non-standard dependency to users isn't great. But CUE /is/
great, and it's my plugin, and I'll do what I want.